Lecture 46 - Bringing FAT16 into 64 bit mode

+++ src/kernel.h
@@ -20,5 +20,5 @@ struct paging_desc* kernel_desc();
 
 #define ERROR(value) (void*)((intptr_t)value)
 #define ERROR_I(value) (int)((intptr_t)(value))
-#define ISERR(value) ((int)value < 0)
+#define ISERR(value) ((int)((int64_t)value) < 0)
 #endif
